+++ py-polars/src/interop/numpy/mod.rs
@@ -1,1 +1,5 @@
-pub mod to_py;
+mod to_py_df;
+mod to_py_series;
+mod utils;
+
+pub(crate) use to_py_series::{series_to_numpy, try_series_to_numpy_view};
